+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- Update permalink in combo result to point to new site's path
+
 # 0.11.0
 
 - Add `autocomplete` method